+++ GymSystem.App/ViewModels/MainViewModel.cs
@@ -1,0 +1,102 @@
+﻿using Microsoft.Toolkit.Uwp.Helpers;
+using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GymSystem.App.ViewModels
+{
+
+
+        /// <summary>
+        /// Provides data and commands accessible to the entire app.  
+        /// </summary>
+        public class MainViewModel : BindableBase
+        {
+            /// <summary>
+            /// Creates a new MainViewModel.
+            /// </summary>
+            public MainViewModel() => Task.Run(GetCustomerListAsync);
+
+            /// <summary>
+            /// The collection of customers in the list. 
+            /// </summary>
+            public ObservableCollection<CustomerViewModel> Customers { get; }
+                = new ObservableCollection<CustomerViewModel>();
+
+            private CustomerViewModel _selectedCustomer;
+
+            /// <summary>
+            /// Gets or sets the selected customer, or null if no customer is selected. 
+            /// </summary>
+            public CustomerViewModel SelectedCustomer
+            {
+                get => _selectedCustomer;
+                set => Set(ref _selectedCustomer, value);
+            }
+
+            private bool _isLoading = false;
+
+            /// <summary>
+            /// Gets or sets a value indicating whether the Customers list is currently being updated. 
+            /// </summary>
+            public bool IsLoading
+            {
+                get => _isLoading;
+                set => Set(ref _isLoading, value);
+            }
+
+            /// <summary>
+            /// Gets the complete list of customers from the database.
+            /// </summary>
+            public async Task GetCustomerListAsync()
+            {
+                await DispatcherHelper.ExecuteOnUIThreadAsync(() => IsLoading = true);
+
+            var customers = await App.Repository.GetCustomers();
+                if (customers == null)
+                {
+                    return;
+                }
+
+                await DispatcherHelper.ExecuteOnUIThreadAsync(() =>
+                {
+                    Customers.Clear();
+                    foreach (var c in customers)
+                    {
+                        Customers.Add(new CustomerViewModel(c));
+                    }
+                    IsLoading = false;
+                });
+            }
+
+        internal  async Task DeletePerson(CustomerViewModel deleteOrder)
+        {
+            await  App.Repository.DeleteCustomer(deleteOrder.Model);
+        }
+
+        /// <summary>
+        /// Saves any modified customers and reloads the customer list from the database.
+        /// </summary>
+        public void Sync()
+            {
+                Task.Run(async () =>
+                {
+                    IsLoading = true;
+                    foreach (var modifiedCustomer in Customers
+                        .Where(customer => customer.IsModified).Select(customer => customer.Model))
+                    {
+                     //   await App.Repository.Customers.UpsertAsync(modifiedCustomer);
+                    }
+
+                    await GetCustomerListAsync();
+                    IsLoading = false;
+                });
+            }
+        }
+    }
+
+
+